% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegCounts.R
\name{createCountDf}
\alias{createCountDf}
\title{Creates a countMatrix corresponding to the tsv files}
\usage{
createCountDf(segNames, tsvFiles, readEnd)
}
\arguments{
\item{segNames}{list containing the segment/segment-pairs corresponding to each tsv file}

\item{tsvFiles}{vector conatining the path of tsvFiles}

\item{readEnd}{character containing read end}
}
\value{
data.frame containing segment counts with rownames as segment/segment-pairs
        columns as cells
}
\description{
Creates a countMatrix corresponding to the tsv files
}
