% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegCounts.R
\name{createSCellObj}
\alias{createSCellObj}
\title{Creates a SingleCellExperiment object given count data and cell annotation file}
\usage{
createSCellObj(counts, cDataPath, sepCData = ",")
}
\arguments{
\item{counts}{data.frame containing segment counts with rownames as segment/segment-pairs
columns as cells}

\item{cDataPath}{character path containing the annotation file corresponding to cells}

\item{sepCData}{character containing the separator that would be used when reading annotation file}
}
\value{
SingleCellExperiment object
}
\description{
Creates a SingleCellExperiment object given count data and cell annotation file
}
