% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegCounts.R
\name{extSegs}
\alias{extSegs}
\title{Gets the combined segment pairs within the tsv files and the type of read end}
\usage{
extSegs(tsvFiles)
}
\arguments{
\item{tsvFiles}{vector conatining the path of tsvFiles}
}
\value{
list with two arguments with first a character containing read end and
        second being another list containing segment/segment-pairs corresponding
        to each tsv file
}
\description{
Gets the combined segment pairs within the tsv files and the type of read end
}
