% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegCounts.R
\name{getTsvFiles}
\alias{getTsvFiles}
\title{Gets the segment count TSV files within a directory}
\usage{
getTsvFiles(dir)
}
\arguments{
\item{dir}{character containing the directory in which tsv files are expected}
}
\value{
Vector of the tsv files with the appropriate path
}
\description{
Gets the segment count TSV files within a directory
}
