% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegCounts.R
\name{readSegCounts}
\alias{readSegCounts}
\title{Reads the segment counts from the directory containing tsv files}
\usage{
readSegCounts(dir, sc = F, cDataPath = NULL, sepCData = ",")
}
\arguments{
\item{dir}{character containing the directory in which tsv files are present}

\item{sc}{boolean denoting whether SingleCellExperiment has to created}

\item{cDataPath}{character denoting path of the cell annotation file}

\item{sepCData}{character denoting the separator to be used for reading
cell annotation file}
}
\value{
data.frame or SingleCellExperiment containing the counts
}
\description{
Reads the segment counts from the directory containing tsv files
}
